Async call on software deploy start
Adding async call on software deploy start for patch. The software
deploy start for major release was already async, making
patch one also async to be used by VIM orchestration

Test Plan:
PASS: Run software deploy start for patch release and
check if its async call
PASS: Verify start-failed case for patch release
PASS: Verify start-done case for patch release
PASS: Create iso and verify software-deploy-patch
script is intalled on right path

Story: 2010676
Task: 50101

+++ software/software/apt_utils.py
@@ -5,11 +5,13 @@ SPDX-License-Identifier: Apache-2.0
 
 """
 import logging
+import os
 import subprocess
 
 from software import constants
 import software.config as cfg
 from software.exceptions import APTOSTreeCommandFail
+from software.exceptions import SoftwareServiceError
 
 LOG = logging.getLogger('main_logger')
 
@@ -87,24 +89,29 @@ def run_install(repo_dir, sw_release, packages):
     :param sw_release: Patch release version (MM.mm.pp)
     :param packages: List of Debian packages
     """
-    try:
-        LOG.info("Running apt-ostree install")
+    LOG.info("Running apt-ostree install")
 
-        packages = " ".join(packages)
-        subprocess.run(
-            ["apt-ostree", "compose", "install",
+    packages = " ".join(packages)
+
+    cmd_path = "/usr/sbin/software-deploy/software-deploy-patch"
+    if not os.path.isfile(cmd_path):
+        msg = "software-deploy-patch script was not found"
+        LOG.error(msg)
+        raise SoftwareServiceError(msg)
+
+    try:
+        subprocess.Popen(
+            ["python", cmd_path,
              "--repo", repo_dir,
              "--branch", "starlingx",
              "--feed", cfg.package_feed,
              "--component", sw_release,
-             packages],
-            check=True,
-            capture_output=True)
-    except subprocess.CalledProcessError as e:
+             "--packages",packages],
+            start_new_session=True)
+        LOG.info("%s subprocess started" % cmd_path)
+    except subprocess.SubprocessError as e:
         msg = "Failed to install packages."
-        info_msg = "\"apt-ostree compose intstall\" error: return code %s , Output: %s" \
-            % (e.returncode, e.stderr.decode("utf-8"))
-        LOG.error(info_msg)
+        LOG.error("Failed to start command: %s. Error %s" % (cmd_path, e))
         raise APTOSTreeCommandFail(msg)
